verify(premerge): split preapproval_gate tests into sibling file for size guard
The added coverage pushed preapproval_gate.rs past the 51200-byte source
size guard. Move the #[cfg(test)] module into preapproval_gate_tests.rs
(referenced via #[path]); both files are now well under the limit.
